fix: move internal helpers out of public exports to pass docs:check
The docs:check CI step requires JSDoc on all exported symbols from
package entrypoints. Moved event collector, usage aggregation, and
output helpers into src/internals.ts so they're importable by tests
but not part of the public API surface.
